requirements.txt
Django==2.1.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
PyJWT==1.6.4
coreapi==2.3.3

